task 6

+++ 0x01-variables_if_else_while/6-print_numberz.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	int a;
+	long a;
 
 	a = 0;
 	while (a <= 9)
